refactor(test): share code
