+++ crates/cli/tests/init.rs
@@ -1,18 +1,9 @@
-use assert_cmd::prelude::*;
+pub mod _utils;
+pub use _utils::*;
+
 use pacquet_testing_utils::{bin::pacquet_with_temp_cwd, fs::get_filenames_in_folder};
 use pretty_assertions::assert_eq;
-use std::{env, ffi::OsStr, fs};
-use tempfile::TempDir;
-
-pub fn exec_pacquet_in_temp_cwd<Args>(args: Args) -> TempDir
-where
-    Args: IntoIterator,
-    Args::Item: AsRef<OsStr>,
-{
-    let (mut command, current_dir) = pacquet_with_temp_cwd();
-    command.args(args).assert().success();
-    current_dir
-}
+use std::{env, fs};
 
 #[test]
 fn should_create_package_json() {
